requirements.txt
flask==2.0.3
google-genai
python-dotenv
SpeechRecognition
pyaudio
pyttsx3
